--- optd-perftest/src/postgres.rs
@@ -1,9 +1,0 @@
-use anyhow::Result;
-
-pub struct PostgresDb {}
-
-impl PostgresDb {
-    pub async fn new() -> Result<Self> {
-        Ok(PostgresDb {})
-    }
-}
